--- src/components/genres-markup.js
@@ -1,7 +1,0 @@
-const createGenresMarkup = (genres) => {
-  return genres.map((genre) => {
-    return `<span class="film-details__genre">${genre}</span>`;
-  }).join(`\n`);
-};
-
-export default createGenresMarkup;
